Make score uint32_t for compat with existing HighScore files

+++ src/Headers/miscscreens.h
@@ -8,7 +8,7 @@
 typedef struct
 {
 	char			name[MAX_NAME_LENGTH+1];
-	unsigned long	score;
+	uint32_t		score;
 }HighScoreType;
 
 #define	DARKEN_PANE_Z	450.0f
